requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.2.4
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2==2.8.3
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.2
waitress==1.3.1
wrapt==1.11.2
django-cors-headers
